+++ ne_version.h
@@ -1,3 +1,3 @@
 ﻿#pragma once
-#define NE_PUBLISHED L"23.05.2026 10:23"
-#define NE_VERSION   L"2026.05.23.10"
+#define NE_PUBLISHED L"23.05.2026 11:18"
+#define NE_VERSION   L"2026.05.23.11"